+++ resources/views/filament/auth-system-info.blade.php
@@ -1,0 +1,3 @@
+<div class="ditib-auth-system-info" style="display: flex; justify-content: flex-end;">
+    <x-system-info align="right" />
+</div>